[design] 03/05 메인페이지 컴포넌트 분리 #182
