+++ app/user/novel/component/MyBook.styled.tsx
@@ -11,14 +11,13 @@ export const MyBookContainer = styled.div`
   scrollbar-color: transparent transparent; 
   min-height: 18.75rem;
 
-
   &::-webkit-scrollbar {
-    height: 4px;
+    height: 0.25rem;
   }
 
   &::-webkit-scrollbar-thumb {
     background-color: rgba(0, 0, 0, 0.2); 
-    border-radius: 10px;
+    border-radius: 0.625rem;
   }
 
   &::-webkit-scrollbar-track {
@@ -26,19 +25,10 @@ export const MyBookContainer = styled.div`
   }
 `;
 
-export const SectionTitle = styled.h2`
-  font-size: 1.2rem;
-  font-weight: bold;
-  margin: 20px 0;
-`
-;
-
-
 export const Card = styled.div`
-  background: #fff;
-  border-radius: 10px;
+  border-radius: 0.625rem;
   padding: 0.625rem;
-  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+  box-shadow: 0 0.125rem 0.3125rem rgba(0, 0, 0, 0.1);
   text-align: center;
   min-width: 7.5rem;
   max-width: 10rem;
@@ -53,30 +43,20 @@ export const Card = styled.div`
   .thumbnail {
     position: relative;
     width: 100%;
-    height: 8.75rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
+    height: auto;
+    border-radius: 0.625rem;
     overflow: hidden;
-  }
-
-  img {
-    max-width: 100%;
-    max-height: 100%;
-    object-fit: cover;
-    border-radius: 10px;
-    border: 1px solid var(--gray-500);
   }
 
   .status {
     position: absolute;
     top: 0.5rem;
     right: 0.5rem;
-    background-color: gray;
-    color: white;
+    background-color: var(--gray-300);
+    color: var(--white-color);
     font-size: 0.75rem;
     padding: 0.15rem 0.5rem;
-    border-radius: 5px;
+    border-radius: 0.3125rem;
   }
 
   .status.complete {
@@ -91,7 +71,6 @@ export const Card = styled.div`
     background-color: var(--leave-color);
   }
 
-  /* 제목은 가운데 정렬 */
   .title {
     text-align: center;
     font-weight: bold;
@@ -102,7 +81,7 @@ export const Card = styled.div`
   .info {
     display: flex;
     flex-direction: column;
-    align-items: flex-start; /* "1단계", "펀딩금액"은 왼쪽 정렬 */
+    align-items: flex-start; 
     justify-content: center;
     width: 100%;
     font-size: 0.9rem;
@@ -112,27 +91,7 @@ export const Card = styled.div`
 
   .info p {
     margin: 0;
-    text-align: left; /* 텍스트 왼쪽 정렬 */
+    text-align: left;
     width: 100%;
-  }
-
-  .progress-container {
-    width: 100%;
-    margin-top: 0.5rem;
-    text-align: center;
-  }
-
-  .progress-bar {
-    width: 100%;
-    height: 0.5rem;
-    border-radius: 5px;
-    background: lightgray;
-    position: relative;
-  }
-
-  .progress {
-    height: 100%;
-    border-radius: 5px;
-    background: navy;
   }
 `;
