+++ app/user/novel/component/SearchBook.styled.tsx
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+export const FilterContainer = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 0.625rem;
+  margin: 0.9375rem 0;`
+;
+
+export const DropdownWrapper = styled.div`
+   width: 25%;
+  `
+;
+
+export const SearchWrapper = styled.div`
+   width: 75%;
+  `
