+++ constants/SEARCH_OPTIONS.ts
@@ -1,0 +1,5 @@
+export const SEARCH_OPTIONS = [
+  { value: "genre", label: "장르" },
+  { value: "title", label: "제목" },
+  { value: "author", label: "작가명" },
+];
